requirements.txt
flask
plaid-python
pre-commit
python-dotenv
